Add an app icon
This icon will be changed in the future, but it
should be very easy to change. Xcode will automatically
generate the required images from the svg in Assets/Original/AppIcon
using imagemagick and inkscape.

This means that you will need imagemagick and inkscape to be installed.
Install imagemagick with `brew install imagemagick`
Install inkscape with `brew cask install inkscape`
This icon also uses the WaifSSK font, which can be found in
Assets/.dependencies/WaifSSK/WaifSSK.ttf